Reimplement NPC setup & code improvements
npc/behaviour_base.py
- moved from npc/npc_behaviour.py

npc/npc.py
- moved global protected variables used only by Player to sprites/player.py
- NPC now inherits from npc.npc_base.NPCBase
- implemented NPC.on_path_abortion

npc/npc_base.py
- extracted from npc.npc.NPC
- provides an abstract interface which NPC now inherits from, and which can be safely referenced by files previously depending on npc.npc.NPC

npc/npc_behaviour.py
- extracted from npc.npc.NPC
- groups all functions regarding NPC specific behaviour

sprites/player.py
- moved global protected variables used only by Player from npc/npc.py

settings.py
- added SCALED_TILE_SIZE constant, which simplifies certain code segments of the NPC pathfinding setup
