+++ src/screens/level.py
@@ -1,11 +1,15 @@
+import math
 import sys
-import pygame 
-
+import pygame
 
 from random import randint
+
+import pytmx
 from pathfinding.core.grid import Grid as PF_Grid
 from pathfinding.finder.a_star import AStarFinder as PF_AStarFinder
 
+from src.npc.npc import NPC
+from src.npc.npc_behaviour import NPCBehaviourMethods
 from src.support import map_coords_to_tile, load_data, resource_path
 from src.groups import AllSprites
 from src.overlay.soil import SoilLayer
@@ -34,8 +38,9 @@ class Level:
         self.switch_screen = switch
 
         # pathfinding
+        self.pf_matrix_size = ()
         self.pf_matrix = []
-        self.pf_grid: PF_Grid
+        self.pf_grid: PF_Grid | None = None
         self.pf_finder = PF_AStarFinder()
 
         # sprite groups
@@ -57,7 +62,8 @@ class Level:
             self.all_sprites,
             self.collision_sprites,
             tmx_maps['main'],
-            frames["level"])
+            frames["level"],
+        )
 
         # weather
         self.sky = Sky()
@@ -90,22 +96,28 @@ class Level:
         self.shop = ShopMenu(self.player, self.toggle_shop, self.font)
         self.shop_active = False
 
-
     # setup
     def setup(self):
         self.activate_music()
 
-        self.setup_layer_tiles('Lower ground', self.setup_environment)
-        self.setup_layer_tiles('Upper ground', self.setup_environment)
-        self.setup_layer_tiles('Water', self.setup_water)
+        self.pf_matrix_size = (self.tmx_maps["main"].width, self.tmx_maps["main"].height)
+        self.pf_matrix = [[1 for _ in range(self.pf_matrix_size[0])] for _ in range(self.pf_matrix_size[1])]
 
-        self.setup_layer_objects('Collidable objects', self.setup_objects)
-        self.setup_layer_objects('Collisions', self.setup_collisions)
-        self.setup_layer_objects('Interactions', self.setup_interactions)
-        self.setup_layer_objects('Entities', self.setup_entities)
+        self.setup_tile_layer('Lower ground', self.setup_environment)
+        self.setup_tile_layer('Upper ground', self.setup_environment)
+        self.setup_tile_layer('Water', self.setup_water)
 
-    def setup_layer_tiles(self, layer, setup_func):
-        for  x, y, surf in self.tmx_maps['main'].get_layer_by_name(layer).tiles():
+        self.setup_object_layer('Collidable objects', self.setup_collideable_object)
+        self.setup_object_layer('Collisions', self.setup_collision)
+        self.setup_object_layer('Interactions', self.setup_interaction)
+        self.setup_object_layer('Entities', self.setup_entity)
+
+        self.pf_grid = PF_Grid(matrix=self.pf_matrix)
+        NPCBehaviourMethods.init()
+        self.setup_object_layer('NPCs', self.setup_npc)
+
+    def setup_tile_layer(self, layer, setup_func):
+        for x, y, surf in self.tmx_maps['main'].get_layer_by_name(layer).tiles():
             x = x * TILE_SIZE * SCALE_FACTOR
             y = y * TILE_SIZE * SCALE_FACTOR
             pos = (x, y)
@@ -119,14 +131,28 @@ class Level:
         image = self.frames['level']['animations']['water']
         AnimatedSprite(pos, image, self.all_sprites, LAYERS['water'])
 
-    def setup_layer_objects(self, layer, setup_func):
+    def setup_object_layer(self, layer, setup_func):
         for obj in self.tmx_maps['main'].get_layer_by_name(layer):
             x = obj.x * SCALE_FACTOR
             y = obj.y * SCALE_FACTOR
             pos = (x, y)
             setup_func(pos, obj)
 
-    def setup_objects(self, pos, obj):
+    def pf_matrix_setup_collision(self, pos: tuple[float, float], size: tuple[float, float]):
+        """
+        :param pos: Absolute position of collision rect (x, y)
+        :param size: Absolute size of collision rect (width, height)
+        """
+        tile_x = int(pos[0] / TILE_SIZE)
+        tile_y = int(pos[1] / TILE_SIZE)
+        tile_w = math.ceil((pos[0] + size[0]) / TILE_SIZE) - tile_x
+        tile_h = math.ceil((pos[1] + size[1]) / TILE_SIZE) - tile_y
+
+        for w in range(tile_w):
+            for h in range(tile_h):
+                self.pf_matrix[tile_y + h][tile_x + w] = 0
+
+    def setup_collideable_object(self, pos, obj: pytmx.TiledObject):
         image = pygame.transform.scale_by(obj.image, SCALE_FACTOR)
 
         if obj.name == 'Tree':
@@ -137,26 +163,42 @@ class Level:
         else:
             Sprite(pos, image, (self.all_sprites, self.collision_sprites))
 
-    def setup_collisions(self, pos, obj):
+        self.pf_matrix_setup_collision((obj.x, obj.y), (obj.width, obj.height))
+
+    def setup_collision(self, pos, obj):
         size = (obj.width * SCALE_FACTOR, obj.height * SCALE_FACTOR)
         image = pygame.Surface(size)
         Sprite(pos, image, self.collision_sprites)
 
-    def setup_interactions(self, pos, obj):
+        self.pf_matrix_setup_collision((obj.x, obj.y), (obj.width, obj.height))
+
+    def setup_interaction(self, pos, obj):
         size = (obj.width * SCALE_FACTOR, obj.height * SCALE_FACTOR)
         image = pygame.Surface(size)
         Sprite(pos, image, self.interaction_sprites, LAYERS['main'], obj.name)
 
-    def setup_entities(self, pos, obj):
+    def setup_entity(self, pos, obj):
         self.entities[obj.name] = Player(game=self.game,
             pos=pos,
             frames=self.frames['character']['rabbit'],
-            groups=self.all_sprites,
+            groups=(self.all_sprites, self.collision_sprites),
             collision_sprites=self.collision_sprites,
             apply_tool=self.apply_tool,
             interact=self.interact,
             sounds=self.sounds,
             font=self.font
+        )
+
+    def setup_npc(self, pos, obj):
+        self.npcs[obj.name] = NPC(pos=pos,
+                                  frames=self.frames['character']['rabbit'],
+                                  groups=(self.all_sprites, self.collision_sprites),
+                                  collision_sprites=self.collision_sprites,
+                                  apply_tool=self.apply_tool,
+                                  soil_layer=self.soil_layer,
+                                  pf_matrix=self.pf_matrix,
+                                  pf_grid=self.pf_grid,
+                                  pf_finder=self.pf_finder
         )
 
     def get_map_size(self):
@@ -280,7 +322,6 @@ class Level:
             entity.blocked = True
             entity.direction = pygame.Vector2(0, 0)
 
-
     # draw
     def draw_overlay(self):
         current_time = self.sky.get_time()
@@ -291,7 +332,6 @@ class Level:
         self.all_sprites.draw(self.player.rect.center)
         self.draw_overlay()
         self.sky.display(dt)
-
 
     # update
     def update_rain(self):
