+++ src/settings.py
@@ -1,4 +1,3 @@
-import pygame
 import pygame.freetype
 import pytmx
 from src.warnings import *
@@ -12,6 +11,7 @@ SCREEN_WIDTH, SCREEN_HEIGHT = 1280, 720
 TILE_SIZE = 16
 CHAR_TILE_SIZE = 48
 SCALE_FACTOR = 4
+SCALED_TILE_SIZE = TILE_SIZE * SCALE_FACTOR
 TB_LAYER = 11
 
 LAYERS = {
